Plugins: add streamer_tcp
